Move collectNonDeferredFields into own utility file
